Revert "Add tests for ReleaseClient"
This reverts commit dbd4fca307582bbd0221711103ee331d973cb111.
