+++ NGitLab.Tests/TagsTests.cs
@@ -20,6 +20,7 @@ namespace NGitLab.Tests
                 Name = "v0.5",
                 Message = "Test message",
                 Ref = project.DefaultBranch,
+                ReleaseDescription = "Test description",
             });
 
             Assert.IsNotNull(result);
@@ -28,6 +29,31 @@ namespace NGitLab.Tests
 
             tagsClient.Delete("v0.5");
             Assert.IsNull(tagsClient.All.FirstOrDefault(x => string.Equals(x.Name, "v0.5", System.StringComparison.Ordinal)));
+        }
+
+        [Test]
+        public async Task Test_can_create_a_release()
+        {
+            using var context = await GitLabTestContext.CreateAsync();
+            var project = context.CreateProject(initializeWithCommits: true);
+            var tagsClient = context.Client.GetRepository(project.Id).Tags;
+
+            var result = tagsClient.Create(new TagCreate
+            {
+                Name = "0.7",
+                Ref = project.DefaultBranch,
+            });
+
+            var release = tagsClient.CreateRelease("0.7", new ReleaseCreate() { Description = "test" });
+            Assert.That(release.TagName, Is.EqualTo("0.7"));
+            Assert.That(release.Description, Is.EqualTo("test"));
+
+            release = tagsClient.UpdateRelease("0.7", new ReleaseUpdate() { Description = "test edited" });
+            Assert.That(release.TagName, Is.EqualTo("0.7"));
+            Assert.That(release.Description, Is.EqualTo("test edited"));
+
+            tagsClient.Delete("0.7");
+            Assert.IsNull(tagsClient.All.FirstOrDefault(x => string.Equals(x.Name, "0.7", System.StringComparison.Ordinal)));
         }
     }
 }
